auctions_staging/pyproject.toml
[project]
name = "quadrotech"
version = "0.1.0"
description = "QuadroTech platform: auctions + nomenclature"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.27.0",
  "jinja2>=3.1.3",
  "python-multipart>=0.0.9",
  "python-dotenv>=1.0.1",
  "pydantic>=2.6.0",
  "pydantic-settings>=2.2.0",
  "SQLAlchemy>=2.0.27",
  "psycopg2-binary>=2.9.9",
  "loguru>=0.7.2",
  "httpx>=0.27.0",
  "lxml>=5.1.0",
  "beautifulsoup4>=4.12.0",
  "apscheduler>=3.10.4",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
